Move cli_commands.definition to cli.definition for FAB
Fix FAB get_provider_info
