+++ providers/fab/provider.yaml
@@ -254,4 +254,4 @@ auth-managers:
   - airflow.providers.fab.auth_manager.fab_auth_manager.FabAuthManager
 
 cli:
-  - airflow.providers.fab.auth_manager.cli_commands.definition.get_fab_cli_commands
+  - airflow.providers.fab.auth_manager.cli.definition.get_fab_cli_commands
