+++ providers/fab/src/airflow/providers/fab/auth_manager/cli/definition.py
