+++ providers/fab/src/airflow/providers/fab/auth_manager/fab_auth_manager.py
@@ -190,7 +190,7 @@ class FabAuthManager(BaseAuthManager[User]):
     @staticmethod
     def get_cli_commands() -> list[CLICommand]:
         """Vends CLI commands to be included in Airflow CLI."""
-        from airflow.providers.fab.auth_manager.cli_commands.definition import get_fab_cli_commands
+        from airflow.providers.fab.auth_manager.cli.definition import get_fab_cli_commands
 
         return get_fab_cli_commands()
 
@@ -736,9 +736,10 @@ class FabAuthManager(BaseAuthManager[User]):
 def get_parser() -> argparse.ArgumentParser:
     """Generate documentation; used by Sphinx argparse."""
     from airflow.cli.cli_parser import AirflowHelpFormatter, _add_command
+    from airflow.providers.fab.auth_manager.cli.definition import get_fab_cli_commands
 
     parser = DefaultHelpParser(prog="airflow", formatter_class=AirflowHelpFormatter)
     subparsers = parser.add_subparsers(dest="subcommand", metavar="GROUP_OR_COMMAND")
-    for group_command in FabAuthManager.get_cli_commands():
+    for group_command in get_fab_cli_commands():
         _add_command(subparsers, group_command)
     return parser
