+++ providers/fab/tests/unit/fab/auth_manager/cli/test_definition.py
